fix some warnings

+++ src/core/mesh.cpp
@@ -35,13 +35,13 @@ NARUKAMI_BEGIN
         assert(count>0);
         assert((start+count)<=triangles.size());
 
-        uint32_t soa_count = (uint32_t)count/SSE_FLOAT_COUNT + 1;
+        size_t soa_count = (uint32_t)count/SSE_FLOAT_COUNT + 1;
 
         std::vector<Point3f> v0_array;
         std::vector<Vector3f> e1_array;
         std::vector<Vector3f> e2_array;
 
-        for(int i=0;i<soa_count*SSE_FLOAT_COUNT;++i){
+        for(size_t i=0;i<soa_count*SSE_FLOAT_COUNT;++i){
             if(i<count){
                 auto v0 = (*triangles[start+i])[0];
                 auto e1 = (*triangles[start+i])[1]-v0;
@@ -59,7 +59,7 @@ NARUKAMI_BEGIN
         }
         std::vector<SoATriangle> soa_triangles;
           
-        for(int i=0;i<soa_count;++i){
+        for(size_t i=0;i<soa_count;++i){
             SoATriangle triangle;
             triangle.v0=load(&v0_array[i*SSE_FLOAT_COUNT]);
             triangle.e1=load(&e1_array[i*SSE_FLOAT_COUNT]);
